+++ src/main/scala/org/podval/store/NumberedStore.scala
@@ -1,8 +1,8 @@
 package org.podval.store
 
-import org.podval.metadata.Names
+import org.podval.metadata.{Names, Numbered}
 
-trait NumberedStore extends Store, org.podval.metadata.Numbered[NumberedStore]:
+trait NumberedStore extends Store, Numbered[NumberedStore]:
   def oneOf: NumberedStores[NumberedStore]
 
   final override def names: Names = oneOf.number2names(number)
@@ -12,3 +12,13 @@ trait NumberedStore extends Store, org.podval.metadata.Numbered[NumberedStore]:
       (oneOf eq other.asInstanceOf[NumberedStore].oneOf)
 
   override def hashCode: Int = 31 * super.hashCode + System.identityHashCode(oneOf)
+
+  override def compare(that: Numbered[NumberedStore]): Int =
+    that match
+      case other: NumberedStore =>
+        val byParent: Int = Integer.compare(
+          System.identityHashCode(oneOf),
+          System.identityHashCode(other.oneOf)
+        )
+        if byParent != 0 then byParent else this.number - other.number
+      case _ => super.compare(that)